New script update_cvs_root
update_cvs allows to change the CVSROOT in an already
checked out directory. It is useful when the CVS server
is down and you need to use another mirror.

usage: update_cvs_root [options] new_cvsroot dir [dir..]
options:
    -e regex        Exclude dirs matching regexp
    -h              Print this help

The script take a regexp (limited by getopt usage) to exclude
some part of the tree, which is helpful for pkgsrc-wip.