pkgsrc/pbulk-builder: use pkgtools/mksandbox
mksandbox is now in its own package, it should be installed before
creating a sandbox. This new mksandbox mount pkgsrc dir in /tree/pkgsrc
we don't want that as it messes too much with the script: sed it back to
what the user gave us as pkgsrc tree dir.

Move the screen/tmux check later so that the bootstrap mode can be
invoked outside screen/tmux.